+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Latest updates:
++ Re-work the HRRR weather model to use herbie (https://github.com/blaylockbk/Herbie) for weather model access
++ minor bug fixes and unit test updates
+
 ## [0.4.3]
 + Force lat/lon/hgt to float32 so that they line up correctly in stitching
 + Add two stage buffer;